Add day 15 solution.
Solve notes:
* My initial approach was to store everything in a 2d structure
that would be updated on each movement. This was a huge amount of
code and ran into issues in part 2.
* After trying to fix the issues for some time, I went back to
the drawing board and considered only the box positions and the
extents of the grid which could handle part 1 / part 2.
* In retrospect I shouldn't have even tried to stuff everything on
an actual 2d grid...
* The code could be updated further to completely remove the grid
and just use a boundary calculation, but too lazy...
* I made a lot of mistakes on this problem so it ended up taking
quite a while to whittle down. Test data saved the day!